fixes TextArea.getCursorPos() on IE8
the following patch fixes the getCursorPos() to work properly in IE8.
Previously if you had:

abc|
def

versus:

abc
|def

where "|" is the cursor position, the getCursorPos() returned 3. That is wrong
in the second case, where it should return 5 (IE uses \r\n, so the cursor
position jumps by 2 on newline). This patch fixes that (and it also works for
any number of newlines in a row).